bgmock-simulator/requirements.txt
# BGMock Simulator Requirements
# Core UI and Visualization
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3

# Kafka and Message Handling
confluent-kafka==2.3.0
kafka-python==2.0.2

# HTTP and API
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.4.2

# Data Validation and Processing
pydantic-settings==2.0.3

# Async and Concurrency
asyncio-contextmanager==1.0.0

# Utilities
colorama==0.4.6
pyyaml==6.0.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0